refactor: centralize rune amount helpers
